+++ tests/integration/small_text/integrations/pytorch/classifiers/test_kimcnn.py
@@ -328,7 +328,7 @@ class CompileTest(unittest.TestCase):
 
             with patch('torch.__version__', new='2.0.0'), \
                     patch('torch.compile', wraps=torch.compile) as compile_spy:
-                classifier.initialize_kimcnn_model()
+                classifier.initialize()
                 compile_spy.assert_called()
 
     def test_initialize_with_pytorch_geq_v2_and_compile_disabled(self):
@@ -342,7 +342,7 @@ class CompileTest(unittest.TestCase):
 
             with patch('torch.__version__', new='2.0.0'), \
                     patch('torch.compile', wraps=torch.compile) as compile_spy:
-                classifier.initialize_kimcnn_model()
+                classifier.initialize()
                 compile_spy.assert_not_called()
 
     def test_initialize_with_pytorch_lesser_v2_and_compile_enabled(self):
@@ -356,7 +356,7 @@ class CompileTest(unittest.TestCase):
         with patch.object(torch, 'compile', return_value=None):
             with patch('torch.__version__', new='1.9.0'), \
                     patch('torch.compile', wraps=torch.compile) as compile_spy:
-                classifier.initialize_kimcnn_model()
+                classifier.initialize()
                 compile_spy.assert_not_called()
 
     def test_initialize_with_pytorch_lesser_v2_and_compile_disabled(self):
@@ -370,7 +370,7 @@ class CompileTest(unittest.TestCase):
         with patch.object(torch, 'compile', return_value=None):
             with patch('torch.__version__', new='2.0.0'), \
                     patch('torch.compile', wraps=torch.compile) as compile_spy:
-                classifier.initialize_kimcnn_model()
+                classifier.initialize()
                 compile_spy.assert_not_called()