+++ CHANGELOG.md
@@ -37,7 +37,8 @@ On the other hand, this also allowed us to deal with further issues that contain
 
 - General
   - Moved `split_data()` method from `small_text.data.datasets` to `small_text.data.splits`.
-
+- Classification:
+  - The `initialize()` methods of all PyTorch-classifiers (KimCNN, TransformerBasedClassification, SetFitClassification) are now more unified. ([#57](https://github.com/webis-de/small-text/issues/57))
 - Utils
   - `init_kmeans_plusplus_safe()` now supports weighted kmeans++ initialization for `scikit-learn>=1.3.0`.